+++ include/zephyr/sys/cpu_load.h
@@ -1,5 +1,6 @@
 /*
  * Copyright (c) 2024, Nordic Semiconductor ASA
+ * Copyright (c) 2025 Analog Devices, Inc.
  *
  * SPDX-License-Identifier: Apache-2.0
  */
@@ -9,6 +10,7 @@
 
 #include <stdbool.h>
 #include <stdint.h>
+#include <zephyr/toolchain.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -16,32 +18,59 @@ extern "C" {
 
 /** @defgroup cpu_load CPU load monitor
  *  @ingroup os_services
- *  @brief Module for monitoring CPU Load
+ *  @brief Module for monitoring CPU load.
  *
- *  This module allow monitoring of the CPU load.
+ *  The CPU load is the fraction of time the CPU spends outside of the idle
+ *  state. Two measurement backends are available and selected through Kconfig:
+ *
+ *  - @kconfig{CONFIG_CPU_LOAD_BACKEND_RUNTIME_STATS} derives the load from the
+ *    scheduler per-CPU runtime statistics. It is portable across architectures
+ *    and supports multiple CPUs.
+ *  - @kconfig{CONFIG_CPU_LOAD_BACKEND_IDLE_HOOK} measures the load using the
+ *    architecture idle hooks. It has lower overhead and can use an optional
+ *    hardware counter for higher precision, but is limited to a single CPU.
+ *
+ *  Regardless of the backend, the load is reported in per mille (0...1000).
  *  @{
  */
 
-/** @brief Hook called by the application specific hook on entering CPU idle. */
-void cpu_load_on_enter_idle(void);
+/** @brief Convert a per mille CPU load value to whole percent. */
+#define CPU_LOAD_PERMILLE_TO_PERCENT(permille) ((permille) / 10)
 
-/** @brief Hook called by the application specific hook on exiting CPU idle. */
-void cpu_load_on_exit_idle(void);
+/** @brief Get CPU load for a specific CPU.
+ *
+ * @param cpu_id CPU index for which to get the load. For the idle-hook backend
+ *		 only the single supported CPU is measured and this argument is
+ *		 ignored.
+ * @param reset  Reset the measurement window after reading.
+ *
+ * @retval >=0 CPU load in per mille (0...1000).
+ * @retval <0  Negative errno code on failure.
+ */
+int cpu_load_get_cpu(unsigned int cpu_id, bool reset);
 
-/** @brief Get CPU load.
+/** @brief Get CPU load for the current CPU.
  *
- * CPU load is measured using a timer which tracks amount of time spent in the
- * CPU idle. Since it is a software tracking there is some small overhead.
- * Precision depends on the frequency of the timer in relation to the CPU frequency.
+ * @param reset Reset the measurement window after reading.
  *
- * @param reset Reset the measurement after reading.
- *
- * @retval >=0 Positive number - CPU load in per mille.
- * @retval <0 Negative number - error code.
+ * @retval >=0 CPU load in per mille (0...1000).
+ * @retval <0  Negative errno code on failure.
  */
 int cpu_load_get(bool reset);
 
-/** @brief Control periodic CPU statistics report.
+/** @brief Hook called on entering CPU idle.
+ *
+ * Only used by the idle-hook backend; called from the architecture idle path.
+ */
+void cpu_load_on_enter_idle(void);
+
+/** @brief Hook called on exiting CPU idle.
+ *
+ * Only used by the idle-hook backend; called from the architecture idle path.
+ */
+void cpu_load_on_exit_idle(void);
+
+/** @brief Control periodic CPU load report.
  *
  * Report logging is by default enabled.
  *
@@ -68,6 +97,23 @@ typedef void (*cpu_load_cb_t)(uint8_t percent);
  */
 int cpu_load_cb_reg(cpu_load_cb_t cb, uint8_t threshold_percent);
 
+/** @brief Get CPU load as a percentage for a specific CPU.
+ *
+ * @deprecated Use cpu_load_get_cpu() instead. Note that cpu_load_get_cpu()
+ *	       returns the load in per mille rather than percent.
+ *
+ * @param cpu_id CPU index for which to get the load.
+ *
+ * @retval >=0 CPU load in percent (0...100).
+ * @retval <0  Negative errno code on failure.
+ */
+__deprecated static inline int cpu_load_metric_get(int cpu_id)
+{
+	int load = cpu_load_get_cpu((unsigned int)cpu_id, true);
+
+	return (load < 0) ? load : CPU_LOAD_PERMILLE_TO_PERCENT(load);
+}
+
 /**
  * @}
  */
@@ -75,6 +121,5 @@ int cpu_load_cb_reg(cpu_load_cb_t cb, uint8_t threshold_percent);
 #ifdef __cplusplus
 }
 #endif
-
 
 #endif /* ZEPHYR_INCLUDE_SYS_CPU_LOAD_H_ */
