+++ include/zephyr/sys/cpu_load_metric.h
@@ -6,48 +6,16 @@
 
 /**
  * @file
- * @brief CPU load monitoring.
- * @ingroup subsys_cpu_load
+ * @brief Deprecated CPU load metric header.
+ *
+ * @deprecated Include <zephyr/sys/cpu_load.h> instead. The cpu_load_metric_get()
+ *	       function is kept there as a deprecated wrapper around
+ *	       cpu_load_get_cpu().
  */
 
 #ifndef ZEPHYR_INCLUDE_SYS_CPU_LOAD_METRIC_H_
 #define ZEPHYR_INCLUDE_SYS_CPU_LOAD_METRIC_H_
 
-#include <stdint.h>
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-/**
- * @brief CPU Load Monitoring
- * @defgroup subsys_cpu_load CPU Load
- * @since 4.3
- * @version 0.1.0
- * @ingroup os_services
- * @{
- */
-
-/**
- * @brief Get the CPU load as a percentage.
- *
- * Return the percent that the CPU has spent in the active (non-idle) state between calls to this
- * function.
- *
- * @param cpu_id The ID of the CPU for which to get the load.
- *
- * @return CPU load in percent (0-100) in case of success
- * @retval -errno code in case of failure.
- *
- */
-int cpu_load_metric_get(int cpu_id);
-
-/**
- * @}
- */
-
-#ifdef __cplusplus
-}
-#endif
+#include <zephyr/sys/cpu_load.h>
 
 #endif /* ZEPHYR_INCLUDE_SYS_CPU_LOAD_METRIC_H_ */
