--- lib/os/cpu_load_metric/CMakeLists.txt
@@ -1,8 +1,0 @@
-# Copyright (c) 2025 Analog Devices, Inc.
-#
-# SPDX-License-Identifier: Apache-2.0
-
-zephyr_include_directories(${CMAKE_CURRENT_SOURCE_DIR})
-
-zephyr_library()
-zephyr_library_sources(cpu_load_metric.c)
