--- lib/os/cpu_load_metric/Kconfig
@@ -1,21 +1,0 @@
-# Copyright (c) 2025 Analog Devices, Inc.
-#
-# SPDX-License-Identifier: Apache-2.0
-
-menuconfig CPU_LOAD_METRIC
-	bool "CPU Load Metric"
-	select EXPERIMENTAL
-	select THREAD_RUNTIME_STATS
-	select SCHED_THREAD_USAGE
-	select SCHED_THREAD_USAGE_ALL
-	select SCHED_THREAD_USAGE_AUTO_ENABLE
-	help
-	  Tracking of CPU load statistics.
-
-if CPU_LOAD_METRIC
-
-module = CPU_LOAD_METRIC
-module-str = CPU Load Metric
-source "subsys/logging/Kconfig.template.log_config"
-
-endif # CPU_LOAD_METRIC
