+++ samples/subsys/cpu_freq/on_demand/prj.conf
@@ -7,6 +7,6 @@ CONFIG_LOG=y
 # CPU Frequency Subsystem Configuration
 CONFIG_CPU_FREQ=y
 CONFIG_CPU_FREQ_LOG_LEVEL_DBG=y
-CONFIG_CPU_LOAD_METRIC_LOG_LEVEL_DBG=y
+CONFIG_CPU_LOAD_LOG_LEVEL_DBG=y
 CONFIG_CPU_FREQ_POLICY_ON_DEMAND=y
 CONFIG_CPU_FREQ_INTERVAL_MS=500
