+++ samples/subsys/cpu_freq/on_demand/tests.yaml
@@ -8,8 +8,8 @@ common:
     type: multi_line
     regex:
       - "^.*<inf> cpu_freq_on_demand_sample: Starting CPU Freq on-demand policy Sample!"
-      - "^.*<dbg> cpu_load_metric.cpu_load_metric_get: CPU[0-9]+ Execution cycles: \
-         [0-9]+, Total cycles: [0-9]+"
+      - "^.*<dbg> cpu_load.cpu_load_get_cpu: CPU[0-9]+ active cycles: \
+         [0-9]+, wall cycles: [0-9]+"
       - "^.*<dbg> cpu_freq_policy_on_demand.cpu_freq_policy_select_pstate: \
          CPU[0-9]+ Load: [0-9]+%"
       - "^.*<dbg> cpu_freq_policy_on_demand.cpu_freq_policy_select_pstate: On-Demand Policy: \
