+++ tests/subsys/cpu_freq/policies/on_demand/prj.conf
@@ -7,7 +7,7 @@ CONFIG_ZTEST=y
 
 CONFIG_CPU_FREQ=y
 CONFIG_CPU_FREQ_LOG_LEVEL_DBG=y
-CONFIG_CPU_LOAD_METRIC_LOG_LEVEL_DBG=y
+CONFIG_CPU_LOAD_LOG_LEVEL_DBG=y
 CONFIG_CPU_FREQ_POLICY_ON_DEMAND=y
 # Long interval so test can run without automatic frequency changes
 CONFIG_CPU_FREQ_INTERVAL_MS=1000000
